Date is on the right now

+++ app/src/main/java/com/example/smartcart/LastShoppingActivity.java
@@ -85,7 +85,7 @@ public class LastShoppingActivity extends AppCompatActivity implements recyclevi
                             int total = 0;
                             for (int i = 0; i < my_items.length; i++) {
                                 String item[] = my_items[i].split(",");
-                                total += Integer.parseInt(item[4]);
+                                total += Double.parseDouble(item[4]);
                                 item tmpItem = new item(item[1], item[0], Double.parseDouble(item[3]), Integer.parseInt(item[2]));
                                 tmp_items.add(tmpItem);
                                                             }
